refactor: split into modules app.js
