+++ components/deleteTask.js
@@ -1,0 +1,10 @@
+const deleteTask = function() {
+  console.log("Delete Task...");
+  
+  const listItem = this.parentNode;
+  const ul = listItem.parentNode;
+      
+  ul.removeChild(listItem);
+};
+
+export { deleteTask };
