+++ index.html
@@ -59,6 +59,6 @@
       </li>
     </ul>
   </div>
-  <script src="app.js"></script>
+  <script type="module" src="app.js"></script>
 </body>
 </html>
